+++ src/bind.c
@@ -171,9 +171,6 @@ static SEXP as_df_row_impl(SEXP x, enum name_repair_arg name_repair, bool quiet)
 
   x = PROTECT_N(vec_chop(x, R_NilValue), &nprot);
 
-  // Clear list_of class info
-  SET_ATTRIB(x, R_NilValue);
-
   r_poke_names(x, nms);
 
   x = new_data_frame(x, 1);